Atualização JPA DAO

+++ src/main/java/br/cnj/projeto/dao/CasoJudicialDAO.java
@@ -5,12 +5,12 @@ import java.util.List;
 import br.cnj.projeto.models.CasoJudicial;
 
 public interface CasoJudicialDAO {
-    CasoJudicial findById(int id);
+    CasoJudicial findById(Long id);
     List<CasoJudicial> findAll();
     List<CasoJudicial> findAll(int page, int pagSize);
     List<CasoJudicial> findAll(int page, int pagSize, String sortBy, String sortOrder);
     void save(CasoJudicial casoJudicial);    
     void update(CasoJudicial updatedCaso);    
-    void partialUpdate(int id, CasoJudicial casoJudicial);
-    void delete(int id);
+    void partialUpdate(Long id, CasoJudicial casoJudicial);
+    void delete(Long id);
 }
